refactor: moved country display flag and name componets to unique folder
